testing removing skip_cleanup
The problem right now is that JASP stores everything in one release, and has been accumulating screenshots since a while. The correct fix for this is to have date folders, and other tagging. But even so, the size of the release will keep getting bigger, which is not ideal. So we need to figure out ways to make the release smaller and restrained to the latest files. 

This particular commit will probably cause a new release, which doesn't help us. So perhaps one way to do this is to setup a cronjob (perhaps in netlify) to push a file update to jasp every day, which will trigger a fresh release through travis.